requirements.txt
paho-mqtt==1.6.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
flask==3.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
matplotlib==3.7.2